Notifier/requirements.txt
requests
pymongo 
python-telegram-bot
dnspython
